chore: cleanup and minor refactoring
