+++ packages/tradershub/src/components/DemoRealSwitcher/DemoRealSwitcher.tsx
@@ -1,49 +1,19 @@
 import React, { useCallback, useEffect, useRef, useState } from 'react';
 import { twMerge } from 'tailwind-merge';
 import { useOnClickOutside } from 'usehooks-ts';
-import { DemoRealSwitcherLoader, useUIContext } from '@/components';
-import { useRegulationFlags } from '@/hooks';
-import { useActiveTradingAccount, useAuthorize, useTradingAccountsList } from '@deriv/api';
+import { DemoRealSwitcherLoader } from '@/components';
+import { useAccountSwitcher, useRegulationFlags } from '@/hooks';
 import { LabelPairedChevronDownSmRegularIcon } from '@deriv/quill-icons';
 import { Button, Text } from '@deriv-com/ui';
 
-type TAccount = {
-    label: string;
-    value: string;
-};
-
-const accountTypes = [
-    { label: 'Demo', value: 'demo' },
-    { label: 'Real', value: 'real' },
-];
-
 const DemoRealSwitcher = () => {
-    const { data: tradingAccountsList } = useTradingAccountsList();
-    const { data: activeTradingAccount } = useActiveTradingAccount();
-    const { switchAccount } = useAuthorize();
+    const { selected, selectAccount, accountTypes } = useAccountSwitcher();
     const { isSuccess } = useRegulationFlags();
     const [isDropdownOpen, setIsDropdownOpen] = useState(false);
-    const activeAccountType = activeTradingAccount?.is_virtual ? 'demo' : 'real';
-    const activeType = accountTypes.find(account => account.value === activeAccountType);
-    const [selected, setSelected] = useState(activeType);
-    const { label, value } = selected || {};
-    const { setUIState } = useUIContext();
+    const { label, value } = selected ?? {};
 
     const ref = useRef(null);
     useOnClickOutside(ref, () => setIsDropdownOpen(false));
-
-    const firstRealLoginId = tradingAccountsList?.find(acc => !acc.is_virtual)?.loginid;
-
-    const demoLoginId = tradingAccountsList?.find(acc => acc.is_virtual)?.loginid;
-
-    useEffect(() => {
-        if (activeType) {
-            setSelected(activeType);
-            setUIState({
-                accountType: activeAccountType,
-            });
-        }
-    }, [activeAccountType, activeType, setUIState]);
 
     useEffect(() => {
         setIsDropdownOpen(false);
@@ -53,25 +23,13 @@ const DemoRealSwitcher = () => {
         setIsDropdownOpen(prevState => !prevState);
     }, []);
 
-    const selectAccount = (account: TAccount) => {
-        setSelected(account);
-        setUIState({
-            accountType: account.value,
-        });
-
-        const loginId = account.value === 'demo' ? demoLoginId : firstRealLoginId;
-        if (loginId) {
-            switchAccount(loginId);
-        }
-    };
-
     if (!isSuccess) return <DemoRealSwitcherLoader />;
 
     return (
         <div className='relative inline-block w-auto ' ref={ref}>
             <Button
                 className={twMerge(
-                    'cursor-pointer w-full py-2 px-6 border-1 border-solid rounded-xs',
+                    'cursor-pointer w-auto py-2 px-6 border-1 border-solid rounded-xs',
                     value === 'demo' ? 'border-status-light-information' : 'border-status-light-success '
                 )}
                 onClick={toggleDropdown}
@@ -98,32 +56,26 @@ const DemoRealSwitcher = () => {
                 </div>
             </Button>
             {isDropdownOpen && (
-                <div className='absolute z-10 items-center w-full top-28 rounded-xs bg-system-light-primary-background shadow-10'>
+                <div className='absolute z-10 flex flex-col items-center w-full top-28 rounded-xs bg-system-light-primary-background shadow-10'>
                     {accountTypes.map(account => (
-                        <div
+                        <button
                             className={twMerge(
-                                'cursor-pointer hover:bg-system-light-hover-background rounded-xs',
+                                'cursor-pointer hover:bg-system-light-hover-background rounded-xs w-full',
                                 account.value === value && 'bg-system-light-active-background'
                             )}
                             key={account.value}
                             onClick={() => selectAccount(account)}
-                            onKeyDown={event => {
-                                if (event.key === 'Enter') {
-                                    selectAccount(account);
-                                }
-                            }}
-                            role='button'
                         >
                             <Text
                                 align='center'
                                 as='p'
-                                className='px-16 py-6 text-center'
+                                className='py-6'
                                 size='sm'
                                 weight={account.value === value ? 'bold' : 'normal'}
                             >
                                 {account.label}
                             </Text>
-                        </div>
+                        </button>
                     ))}
                 </div>
             )}
