+++ packages/tradershub/src/components/OptionsAndMultipliersSection/OptionsAndMultipliersHeading/OptionsAndMultipliersHeading.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { CurrencySwitcher, StaticLink, TitleDescriptionLoader } from '@/components';
 import { useRegulationFlags } from '@/hooks';
-import { Text, useDevice } from '@deriv-com/ui';
+import { Text } from '@deriv-com/ui';
 
 const getDescription = (isEU: boolean) => {
     if (isEU) {
@@ -27,7 +27,6 @@ const getDescription = (isEU: boolean) => {
  * @returns {React.ElementType} The `OptionsAndMultipliersHeading` component.
  */
 const OptionsAndMultipliersHeading = () => {
-    const { isDesktop } = useDevice();
     const { isSuccess: isRegulationAccessible } = useRegulationFlags();
     const { isEU } = useRegulationFlags();
 
@@ -40,11 +39,11 @@ const OptionsAndMultipliersHeading = () => {
     return (
         <div className='flex flex-col items-start justify-between gap-16 lg:flex-row lg:gap-48'>
             <div className='gap-2 lg:flex lg:flex-col'>
-                {isDesktop && (
+                <div className='d-none lg:block'>
                     <Text size='lg' weight='bold'>
                         {title}
                     </Text>
-                )}
+                </div>
                 {description}
             </div>
             <CurrencySwitcher />
